requirements.txt
discord.py
asyncio
pymongo
dnspython